Append extra exception data to RpcMessage
If there has been any additional data added to an RpcException object, pack
those vars into the RpcMessage buffer that will be sent on the wire.

We always append a kVoid indicator to the end of an exception message, and the
client-side code should look for this to be sure that it's gotten all valid
data and no spurious data.