Unpack extra data from an exception message.
If the server code raises an exception that contains extra var data in its
payload, this unpacks the data and appends it to the RpcException object that
gets raised.

The extra data is not formally structured; it's a list of var objects
(restricted to the types that are supported in RpcMessage.h/cpp) that can be
accessed by index.